added API Key authentication. fixes #3

+++ index.js
@@ -6,6 +6,23 @@ require('dotenv').config()
 const port = process.env.PORT || 3000
 
 const Paste = require('./models/paste')
+const User = require('./models/user')
+
+async function authentication(req, res, next) {
+    const apiKey = req.headers["x-api-key"]
+    if (apiKey == undefined) {
+        res.status(401)
+        res.send('Access Forbidden. No API Key Provided')
+    } else {
+        let response = await User.where('apiKey').equals(apiKey)
+        if (response.length == 0) {
+            res.status(401)
+            res.send('Access Forbidden. Invalid API Key')
+        } else {
+            next()
+        }
+    }
+}
 
 app.use(express.urlencoded({extended: true}))
 
@@ -18,15 +35,29 @@ const db = mongoose.connection
 db.on('error', error => console.log(error));
 db.once('open', () => {console.log('Connected to Mongoose')})
 
-// Routes
+// Create API Key
+app.get('/generateApiKey', async (req,res) => {
+    let response = await User.create({})
+    res.send(response)
+})
 
+// Auth Middleware
+app.use(authentication)
+
+// Routes
 app.get('/pastes', async (req,res) => {
-    let response = await Paste.find({})
+    const apiKey = req.headers["x-api-key"]
+    const userdata = await User.where('apiKey').equals(apiKey)
+    const creatorId = userdata[0].creatorId
+    let response = await Paste.find({'creatorId': creatorId})
     res.send(response)
 })
 
 app.post('/pastes', async (req,res) => {
-    let {message, creatorId} = req.body
+    const apiKey = req.headers["x-api-key"]
+    const userdata = await User.where('apiKey').equals(apiKey).limit(1)
+    const creatorId = userdata[0].creatorId
+    let {message} = req.body
     let response = await Paste.create({
         body: message,
         creatorId
@@ -35,24 +66,51 @@ app.post('/pastes', async (req,res) => {
 })
 
 app.get('/pastes/:pasteId', async (req,res) => {
+    const apiKey = req.headers["x-api-key"]
+    const userdata = await User.where('apiKey').equals(apiKey).limit(1)
+    const creatorId = userdata[0].creatorId
     let {pasteId} = req.params
-    let response = await Paste.findOne({'pasteId':pasteId})
-    res.send(response)
+
+    let response = await Paste.findOne({'pasteId':pasteId, 'creatorId': creatorId})
+    if (response != null) {
+        res.send(response)
+    } else {
+        res.status(403)
+        res.send("Paste doesn't exist or it's private")
+    }
 })
 
 app.patch('/pastes/:pasteId', async (req,res) => {
+    const apiKey = req.headers["x-api-key"]
+    const userdata = await User.where('apiKey').equals(apiKey).limit(1)
+    const creatorId = userdata[0].creatorId
     let {pasteId} = req.params
     let {updatedMessage} = req.body
-    await Paste.findOneAndUpdate({'pasteId': pasteId},
+    let response = await Paste.findOneAndUpdate({'pasteId': pasteId, 'creatorId': creatorId},
         {"$set": {"body": updatedMessage}}
     )
-    res.sendStatus(201)
+    if (response != null) {
+        res.status(201)
+        res.send(response)
+    } else {
+        res.status(403)
+        res.send("Invalid Operation. Either Paste doesn't exist OR You aren't the creator of the Paste")
+    }
 })
 
 app.delete('/pastes/:pasteId', async (req,res) => {
+    const apiKey = req.headers["x-api-key"]
+    const userdata = await User.where('apiKey').equals(apiKey).limit(1)
+    const creatorId = userdata[0].creatorId
     let {pasteId} = req.params
-    let response = await Paste.deleteOne({'pasteId': pasteId})
-    res.send(response)
+    let response = await Paste.deleteOne({'pasteId': pasteId, 'creatorId': creatorId})
+    if (response.deletedCount == 0) {
+        res.status(403)
+        res.send("Invalid Operation. Either Paste doesn't exist OR You aren't the creator of the Paste")
+    } else {
+        res.status(200)
+        res.send('Paste deleted successfully')
+    }
 })
 
 // listen to incoming connections
